web_nlp_module/app/model_server/requirements.txt
spacy
nltk
newspaper3k
lxml-html-clean
fastapi
pydantic
uvicorn[standard]
requests
beautifulsoup4
lxml
stealth-requests
accelerate
bitsandbytes
litserve
transformers